move consistency test to base test
